Correct resident tour model
Closes #8
The problem was caused by an inconsistency in the resident mode
choice and stop choice models.  A walk tour could be created if the
total distance was less than three miles, but if either leg of the trip
was greater than 1.5 miles, the stop location model would crash.
The SLC UEC (the xls) was updated to make it consistent.