Change approach to env variables
Due to limitations with git, the repo should not have the actual
environment variable file (CTRampEnv.bat), because they can be
different on each machine.  Instead, the example file is included in
the repository.  The README directs the user to copy/rename that
to CTRampEnv.bat and then configure it to their local settings.  The
.gitignore file should ignore that new (.bat) file.