Cap 2045 person age at 64
The se data is the same as before excep the max person age is set to 64.
This adjustment was done in an attempt to capture the improved mobility
that seniors will have in the future given the growing prevalence of TNCs.
This change led to a small uptick in VMT/VHT when tested.